% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{create_conda_env}
\alias{create_conda_env}
\title{Make a new conda environment}
\usage{
create_conda_env(condaenv, version = NULL, verbose = TRUE)
}
\arguments{
\item{condaenv}{Specify conda environment name}

\item{version}{Set path to specific version of python.}

\item{verbose}{TRUE or FALSE. When TRUE, shows python and conda configuration.
Default: TRUE}
}
\description{
This function creates a new conda environment and initializes
the new conda environment.
In doing so, this function sets your python version 
and one may specify a specific python version. This is useful
if you have multiple versions of python installed.
When making a new conda environment, if the python version isn't set, then
your default one will be used.
}
\examples{
\dontrun{
create_conda_env( 
    condaenv = "r-reticulate",
    version = "~/anaconda3/bin/python",
    verbose = TRUE)
}
}
