% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lgbm_models.R
\name{run_lgbm}
\alias{run_lgbm}
\title{Run LGBM Model}
\usage{
run_lgbm(
  train_matrix,
  train_matrix_ret,
  test_matrix,
  lgb_params_1,
  lgb_params_2
)
}
\arguments{
\item{train_matrix}{training matrix}

\item{train_matrix_ret}{training matrix of returned users}

\item{test_matrix}{test matrix}

\item{lgb_params_1}{logistic regression params}

\item{lgb_params_2}{regression params}
}
\value{
predictions and feature importance
}
\description{
Run LGBM Model
}
