% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{set_python_virtual}
\alias{set_python_virtual}
\title{Set your Python Virtual environment}
\usage{
set_python_virtual(virtualenv, verbose = TRUE)
}
\arguments{
\item{virtualenv}{Specify virtual environment name}

\item{verbose}{TRUE or FALSE. When TRUE, shows python and virtual environment configuration.
Default: TRUE}
}
\description{
This function sets your virtual environment.
Run this command before PreprocessMatrix. Install python dependencies in the
same virtual environment that you set here.
To make a new virtial environment use the create_virtual_env function.
}
\examples{
\dontrun{
set_python_virtual(
    virtualenv = "r-reticulate", 
    verbose = TRUE)
}
}
